requirements.txt
Django==5.2.6
python-decouple==3.8
django-crispy-forms==2.0
crispy-bootstrap5==0.7
gunicorn==20.1.0
whitenoise==6.4.0
dj-database-url==2.0.0
psycopg2-binary>=2.9.6
Pillow==10.4.0
django-cloudinary-storage==0.3.0
cloudinary==1.30.0
django-allauth==65.11.2
PyJWT>=2.8.0
cryptography>=41.0.4
requests>=2.31.0
africastalking>=2.0.0
django-extensions==3.2